Split dobj from dasm.
